split AuthCont into AuthProv to fix fastRefresh
